This file is not part of the source code
I agree that it is handy to have, but it should not be included in the actual source code. When the program is run it is possible to select any .csv file, and the program is not actually dependent on this file existing. If it is, then it needs to be fixed.